+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Added
 
+- Organization Methods ([#3](http://github.com/lazoliver/imob-backend/pull/3))
+
+### Added
+
 - Logs and Vars ([#2](http://github.com/lazoliver/imob-backend/pull/2))
 
 ### Added